Be more careful about the ordering of connections.
We previously opened both Tor and client connections at once, and
registered all callbacks immediately. This means that one connection
could get a data callback and try to write to the other socket before it
was even connected. Now the Tor connection is made first, then the
client connection, and only then are the data callbacks registered.