Put a buffer on local connections.
We buffer data (up to a limit) before a local is matched with a remote.
After matching, the buffered data is sent to the remote.

This solves a bug where a local could connect and disconnect before
being matched with a remote, and the local socket would stick around
forever.