Store references to proxy pairs to avoid garbage collection.
Adding an event listener is not enough to keep a proxy pair alive. Pairs
could be garbage-collected at unpredictable times. We were able to
provoke this by running more than one pair at a time with servers on
localhost; perhaps it was the greater memory pressure (64 KB blocks at a
time) that exposed it.