Make facilitator requests block until a reg is available.
When a GET request happens, no longer return an empty string when there
are no registrations to hand out. Instead, let the HTTP connection hang
until there is something, then return. It can happen that the HTTP
client (a flash proxy) disconnects before receiving its response; if we
fail to send the reg that has just become available, add it back to the
front of the queue to be handed out again.