Revert facilitator blocking until a reg is available.
This reverts
91cceeef62061d66732be7039d559f5fff49e8ef Document the facilitator's waiting until a reg is available.
c3ea89cb2a2c1e5f8496ec526b9d1c7451e2d06d Log the socket error when we fail to send a reg.
90510d7bf565f0cc9876665aa6e85b55c8e5a3c1 Make facilitator requests block until a reg is available.

This would be a nice model, but the URLLoader in ActionScript doesn't
work exactly the way I expect it to (at least with Flash Player in
Firefox), and it means that this "long polling" technique doesn't work
robustly.

The first time swfcat is loaded, everything is fine: it makes the
/crossdomain.xml request, then starts a URLLoader for /. If there is no
client to be served immediately, then the URLLoader just hangs until
there is one (this part works well).

What goes wrong is when you refresh the page (or navigate away and then
back again). I expected that the URLLoader would break its connection
when this happens, which would be noticed by the facilitator. Instead,
what seems to happen is the URLLoader's HTTP connection goes into limbo,
still alive and waiting for a response, but not hooked up to anything on
the client side. (I suppose this is an artifact of Flash Player using
the browser to make the request, so it might differ on other browsers.)

The new instance of swfcat asks for /crossdomain.xml, and claims to be
opening a URLLoader for /, but instead what appears to happen is the new
request gets queued behind the pending one that is in limbo; tcpdump
doesn't see any second request until the first is completed. Now, if a
client registers, its reg goes out over the in-limbo connection and
effectively disappears. The queued URLLoader request from the proxy gets
sent, but at this point there is no more client reg because it was
presumed to be handled by the first proxy request.

This queueing behavior is evident if you refresh the page, say, five
times, and then kill the facilitator: you see one HTTP request, then a
RST from the server, then four other immediate SYN-RST pairs.

So instead, we go back to the facilitator always returning immediately
(possibly without a client), and the proxy polling on a regular basis
for clients.