Keep retrying remote-tor downloads.
Sometimes a proxy going down at just the right moment can cause Tor to
reject a SOCKS connection (with code 91). This seems to confuse Wget's
automatic retrying; it starts over from the beginning. I have also seen
it abort the Wget download completely. So download the file to an actual
file rather than /dev/null, and run Wget with -c in a loop.