Document that the experiments require Firefox 8.
Firefox 9 changed the meaning of the -no-remote option. Previously, it
meant "don't connect to any already running Firefox, but start a new
instance." Now it additionally means "don't accept any future -remote
commands in this new instance." This is broken for the experiments,
which need the first behavior in order to start multiple instances with
different profiles, but can't have the second behavior because the
instances need to be driven by -remote.

This is the bug that led to the change in the meaning of -no-remote:
"-no-remote browser accepts remote commands"
https://bugzilla.mozilla.org/show_bug.cgi?id=650078

There are other bugs that propose to restore some way to get the old
behavior in a future version.
"split -new-instance flag out of existing -no-remote flag"
https://bugzilla.mozilla.org/show_bug.cgi?id=716110

Another related bug number:
"Firefox 9 Fails To Correctly Use -new-window"
https://bugzilla.mozilla.org/show_bug.cgi?id=703021