Use plain http (not https) in embed links.
Using https would be preferable, which is why I changed this to explicit
https in 836b7a27. But I learned that Firefox doesn't allow making a
plaintext (ws) WebSocket connection in an https page; you have to use
wss instead, for the same reason you get mixed-content warnings.

https://bugzilla.mozilla.org/show_bug.cgi?id=662692
https://developer.mozilla.org/en/Firefox_8_for_developers#WebSockets
https://developer.mozilla.org/en/WebSockets/Writing_WebSocket_client_applications#Security_considerations

I think that we can't use wss sockets because they would require each
client to have a CA cert.

In the iframe link, I've left the protocol implicit, so it will use
https if the page itself is https and http otherwise. This is to avoid
using mixed content, but the badge won't work if the browser disallows
the plaintext WebSocket.

The about:config setting that controls this is
network.websocket.allowInsecureFromHTTPS.