Add "LearnCircuitBuildTimeout 0" to torrc.
This improves reliability when reconnecting over flash proxies; the
rational is given in a1bf13ae.

I am adding an explicit "CircuitBuildTimeout 60", even though tor(1)
says that is the default, because without it I get the message

Jul 04 14:08:37.683 [warn] CircuitBuildTimeout is shorter (0 seconds) than recommended (10 seconds), and LearnCircuitBuildTimeout is disabled.  If tor isn't working, raise this value or enable LearnCircuitBuildTimeout.