Add Drupal module by Brian Duggan.
Brian did research and found out what files and syntax are needed for a
Drupal 7 module. This module loads and runs the flash proxy program, but
it doesn't put a badge on the page.

The module originally had a copy of flashproxy.js with the only change
being an additional line at the bottom to start the proxy running. I
changed that into a symlink to the main flashproxy.js, and added a new
flashproxy-start.js that calls the start method. I haven't tested these
changes at all or even know if I have the syntax right.