Don't let websockify send "Sec-WebSocket-Protocol: binary".
We patch websockify so that it assumes binary transport when no
subprotocol is given. Previously we continued to send a
"Sec-WebSocket-Protocol: binary" in our response, which is a violation
of section 4.2.2 of RFC 6455:
	"The value chosen MUST be derived from the client's handshake,
	specifically by selecting one of the values from the
	|Sec-WebSocket-Protocol| field that the server is willing to use
	for this connection (if any).  If the client's handshake did not
	contain such a header field or if the server does not agree to
	any of the client's requested subprotocols, the only acceptable
	value is null."
Chromium (version 20, I believe) started being more strict about
enforcing this, which caused the relay connection to fail.

https://bugs.webkit.org/show_bug.cgi?id=82307
http://trac.webkit.org/changeset/112499