Listen on both IPv4 and IPv6 by default in flashproxy-client.
If no host is given for the local address, listen on 127.0.0.1 and ::1.
If no host is given for the remote address, listen on 0.0.0.0 and ::.

This adds the ability to listen on more than one local or remote socket,
though this is not exposed in the command-line interface except in this
special case.