Add a separate --register-addr option taking an argument.
Use this when the address by which you are reached from the outside
world is not the same one as the remote addr. Normally when registering,
you just leave off the host and allow the facilitator to fill it in
based on the address the registration is coming from. But this fails
when, for example, your registration is made over HTTP over IPv6, but
you can only accept connections over IPv4. You could also use this to
listen on a specific internal LAN address, while registering your public
NAT IP and a (presumably forwarded) port.