Capture registration helper stdout, stderr, and exit status.
The registration helpers run in a thread, waiting in an infinite loop
until the main process asks for a registration. This is necessary now
because flashproxy-client needs to register multiple times in order to
keep getting service. When #5426 is done, it should only be necessary to
register once in the common case, and the thread won't be necessary.