Start the registration thread only after background forking.
I don't know the exact cause, but something was causing the registration
thread to hand at the subprocess.Popen call, but only when the program
was run with the --daemon option, which causes it to fork at startup.
Starting the thread after the fork seems to make the problem go away. It
might be this bugs I found while searching:
	http://bugs.python.org/issue13817