Added asciidoc file for flashproxy-reg-email man page.
Added the asciidoc source for the flashproxy-reg-email
man page. The title on this one is a bit big, with the
default options a2x will truncate it. To generate a manpage
with the full title use:

a2x --no-xmllint --xsltproc-opts "--stringparam man.th.title.max.length 23" \
-d manpage -f manpage flashproxy-reg-email.1.txt

The style is based on the man pages for tor and asciidoc.