Add a copy of the GPL, copied from pyinstaller-2.0.
I think that we are not actually bound to include the GPL in the
distribution just because we use PyInstaller; its license has a blanket
exception for linking with an unmodified copy of the bootloader. See
http://www.pyinstaller.org/export/v2.0/project/doc/Manual.html#license
http://www.pyinstaller.org/wiki/FAQ#License
I'm including it because the copyright text refers to it, and because it
doesn't hurt anything other than package size.