Do not set test cookie to determine if cookies are enabled.
If navigator.cookieEnabled is undefined, the current code will
fall back to setting a test cookie to determine if cookies
are enabled. This may overwrite an already present cookie,
and may cause trouble for people who wish to host the badge
on their own websites. This drops support for older browsers
such as IE <= 4 and Netscape <= 6 (each more than 10 years old).