Add text to options page when flashproxy cookie is not present.
If the user visits the page with no cookie set, some text informs
him that he has not chosen a setting yet, and that his browser may
be used as a proxy depending on how the badge has been configured by
the website admin. The "configuration" in question is whether the badge is hosted
with the "cookierequired" query parameter or not.