Handle opt-in cookie in flashproxy.js
The proxy disables itself if either the user has opted out
or opt-in is required and the user has not opted-in. To
determine if opt-in is required it checks for a "cookierequired"
boolean query parameter. To determine the user's opt-in setting
it parses document.cookie and interprets the value of the
cookie as a boolean parameter.