Make facilitator_poll_interval option override other settings.
If the facilitator_poll_interval option is provided, and is a valid
value, it overrides the facilitator-provided interval and the default
interval. If the option is missing, fall back on the facilitator
interval, and absent that, the default interval. If any value is smaller
than MIN_FACILITATOR_POLL_INTERVAL, that is used instead.