Fix a race condition in registration.
This could cause the first registration message not to be sent. The
registration thread used the "notify" of a condition variable as the
signal to do a registration. If the main thread did its "notify" before
the registration thread did its "wait", then the registration thread
would not wake up until the next time a registration was requested.