Use resolved listener addresses for port forwarding.
Previously we used remote_addr, the actual address given on the command
line. The difference is that after opening the listener, we know what
port ":0" became, so this is now possible:
	./flashproxy-client --external --port-forwarding :9001 :0
The other thing we can do is skip IPv6 addresses, because it appears
that UPnP and NAT-PMP are for TCP/IPv4 ports only.

The above command will produce output like

2014-02-01 17:07:01 Listening remote on 0.0.0.0:40138.
2014-02-01 17:07:01 Listening remote on [::]:55179.
2014-02-01 17:07:01 Listening local on 127.0.0.1:9001.
2014-02-01 17:07:01 Listening local on [::1]:9001.
2014-02-01 17:07:01 Not forwarding to [::]:55179 because it is not an IPv4 address.
2014-02-01 17:07:01 Running port forwarding command: tor-fw-helper -p 40138:40138