Add "DOSLogDir /dev/null" to the example Apache configuration.
By default, mod_evasive creates a file in /tmp for every blocked client,
with a name like dos-<IP address>. I didn't know this was going on, and
found 1,659 such files on the facilitator dating back to May 2013. The
contents of each file is just a PID, I think.

This configuration option prevents the creation of such files. Blocked
IP addresses still turn up, however, in the error.log that is not
disabled, in lines that look like
	[DATE] [error] [client X.X.X.X] client denied by server configuration: /usr/local/bin/facilitator.cgi